pyproject.toml
[project]
name = "dustboy"
version = "0.1.0"
description = ""
authors = [
    {name = "suthinan",email = "suthinan.rpn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-mongoengine-3 (>=1.1.0,<2.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dustboy-web = "dustboy.cmd.web:main"